+++ a2asrv/agentexec.go
@@ -196,6 +196,7 @@ func (f *factory) loadExecutionContext(ctx context.Context, tid a2a.TaskID, para
 		task: &taskupdate.VersionedTask{
 			Task:    storedTask,
 			Version: lastVersion,
+			Stored:  true,
 		},
 		reqCtx: &RequestContext{
 			Message:    msg,
@@ -246,7 +247,7 @@ func (f *factory) CreateCanceler(ctx context.Context, params *a2a.TaskIDParams) 
 	}
 
 	canceler := &canceler{agent: f.agent, reqCtx: reqCtx, task: task, interceptors: f.interceptors}
-	updateManager := taskupdate.NewManager(f.taskStore, &taskupdate.VersionedTask{Task: task, Version: version})
+	updateManager := taskupdate.NewManager(f.taskStore, &taskupdate.VersionedTask{Task: task, Version: version, Stored: true})
 	processor := newProcessor(updateManager, f.pushConfigStore, f.pushSender, reqCtx, f.taskStore)
 	return canceler, processor, &cleaner{agent: f.agent, reqCtx: reqCtx}, nil
 }
@@ -315,8 +316,6 @@ type processor struct {
 	pushSender      PushSender
 	reqCtx          *RequestContext
 	store           TaskStore
-
-	processedCount int
 }
 
 var _ taskexec.Processor = (*processor)(nil)
@@ -381,11 +380,6 @@ func (p *processor) Process(ctx context.Context, event a2a.Event) (*taskexec.Pro
 // ProcessError implements taskexec.ProcessError interface method.
 // Here we can try handling producer or queue error by moving the task to failed state and making it the execution result.
 func (p *processor) ProcessError(ctx context.Context, cause error) (a2a.SendMessageResult, error) {
-	if p.reqCtx.StoredTask == nil && p.processedCount == 0 {
-		// there was no task in the store, don't create one in failed state and allow to error to be propagated to the client
-		return nil, cause
-	}
-
 	versioned, err := p.updateManager.SetTaskFailed(ctx, nil, cause)
 	if err != nil {
 		return nil, err